requirements.txt
alembic==1.11.1
anyio==3.6.2
aws-psycopg2==1.3.8
boto3==1.26.146
botocore==1.29.146
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click
cognitojwt==1.4.1
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.2
fastapi-cognito==2.3.0
greenlet==2.0.2
h11
idna==3.4
jmespath==1.0.1
keras==2.12.0
Mako==1.2.4
mangum==0.17.0
MarkupSafe==2.1.2
numpy
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.30.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.27.0
typing_extensions==4.5.0
urllib3==1.26.16
uvicorn